handle subprocess cwd better
When GladTeX was single-threaded, it used to change the cwd so that LaTeX would
put the temporary files along with the output file in the same directory. Now
the cwd is passed to the new child (well, it changes the directory after the
fork), because multiple parallel threads changing the cwd was disastrous.