Strip formatting from alt attribute
This commit adds a small, yet helpful feature to the text of alt
attributes of the generated images. Reading LaTeX is not as convenient
as reading the formula in the image, yet blind people or those with
images turned of have to do so. Some formatting commands as `\,` or
`\left` or `\big` do not influence the meaning of the formula. These are
stripped now, so that

    a\,b\;c =\ d

becomes

    a b c = d

In the alt text. The image is still rendered with the complete formula.