Fix image cache sharing between multiple documents
Previously, the search for the next image file name to use for a new
cache entry did not look into the correct directory and so always did
not find an existing cached formula image which led to restarting with
`eqn000` on every invocation, essentially overwriting the cache. This
should now be fixed.