Try to detect non-nestable math environments
Currently, all formulas are wrapped in a standard math environment
(inline math or display math), but as there are some math environments
for more advanced formatting which cannot be nested in such
environments, they could not be used so far.

To change that, we try to detect those environments following a simple
heuristic: If the formula begins immediately with an environment
(optionally prefixed by any amount of whitespace and comments) and this
environment is in a list of non-nestable math environments, then this
formula is not wrapped in a standard environment. This probably does not
cover all possible use cases, but provides a simple backwards compatible
solution that should support the most common ones.

Fixes #21.