Optimize non-nestable math environment detection regex
The previously used regular expression should be correct, but during
testing it appeared, that in some (possibly real-world) cases with many
empty lines after a comment or many spaces before the environment
beginning it was terribly, unacceptably slow. It has now been optimized
and even made minimally more correct to mitigate that, so it hopefully
should be usable now.