Support embedding of excluded formulas in the document
When the new command line option `-I`/`--embed-excluded-formulas` (or
`--epub` which implies it) is passed, formulas too long for an image
`alt` attribute are appended at the end of the document in an `<aside>`
with a heading (which can be specified by the new `-H` option) such that
the converted image displaying the formula links to that in the
`<aside>` and the formula in the `<aside>` links back to the image,
instead of writing them to an extra file. Both acting as Pandoc filter
and editing an HTML file directly are supported.

Closes #15.