Allow passing `-r` option without explicitly ignoring `-f` (#35)
Previously, passing `-r` on the command line without passing an
equivalent of `-f ''` along as well to disable the `-f` option would
result in an error, despite the options being mutually exclusive. This
was caused by always storing the default constant for `-f` even when it
wasn't passed, so it was not possible anymore to distinguish whether the
option was passed or not. Now the default constant is only saved to the
parsed option value iff neither `-f` nor `-r` have been passed on the
command line.

Fixes #35.