Check Pandoc AST version and refactor `"blocks"` parsing
When extracting the AST blocks of Pandoc JSON AST root, the API version
is now checked for compatibility; as the version does not adhere to
semantic versioning and breaks between minor versions, but only rarely
changes at all, we only support the latest minor version.

All accesses to these root AST blocks are now also refactored to be
handled in the `ast` module, finally completing the delegation of the AST
handling to it.